# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NuclearLedger API",
    "description": "NuclearLedger 记账软件后端接口文档",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "本地开发环境"
    }
  ],
  "tags": [
    { "name": "基础", "description": "健康检查" },
    { "name": "认证", "description": "注册登录" },
    { "name": "用户", "description": "用户信息" },
    { "name": "记账记录", "description": "收支记录 CRUD" },
    { "name": "分类", "description": "收支分类 CRUD" },
    { "name": "账户", "description": "账户 CRUD" },
    { "name": "预算", "description": "预算 CRUD" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "message": { "type": "string", "example": "success" },
          "data": {}
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "example": "test", "description": "用户名" },
          "password": { "type": "string", "example": "123456", "description": "密码，最少6位" },
          "nickname": { "type": "string", "example": "Test User", "description": "昵称，可选" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "example": "test" },
          "password": { "type": "string", "example": "123456" }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "username": { "type": "string", "example": "test" },
          "nickname": { "type": "string", "example": "Test User" },
          "email": { "type": "string", "nullable": true },
          "avatarUrl": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "example": "2026-06-03 14:30:00" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "example": "eyJhbGciOiJIUzM4NCJ9..." },
          "tokenType": { "type": "string", "example": "Bearer" },
          "user": { "$ref": "#/components/schemas/UserResponse" }
        }
      },
      "TransactionCreateRequest": {
        "type": "object",
        "required": ["accountId", "categoryId", "amount", "type", "transactionDate"],
        "properties": {
          "accountId": { "type": "integer", "example": 1, "description": "账户ID" },
          "categoryId": { "type": "integer", "example": 1, "description": "分类ID" },
          "amount": { "type": "number", "example": 25.50, "description": "金额，必须>0" },
          "type": { "type": "string", "enum": ["expense", "income"], "example": "expense", "description": "类型" },
          "note": { "type": "string", "example": "午餐", "description": "备注，可选" },
          "transactionDate": { "type": "string", "example": "2026-06-03 14:30:00", "description": "记账时间" }
        }
      },
      "TransactionResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "accountId": { "type": "integer", "example": 1 },
          "categoryId": { "type": "integer", "example": 1 },
          "categoryName": { "type": "string", "example": "餐饮" },
          "accountName": { "type": "string", "example": "现金" },
          "amount": { "type": "number", "example": 25.50 },
          "type": { "type": "string", "example": "expense" },
          "note": { "type": "string", "example": "午餐" },
          "transactionDate": { "type": "string", "example": "2026-06-03 14:30:00" },
          "createdAt": { "type": "string", "example": "2026-06-03 14:31:00" },
          "updatedAt": { "type": "string", "example": "2026-06-03 14:31:00" }
        }
      },
      "CategoryCreateRequest": {
        "type": "object",
        "required": ["name", "type"],
        "properties": {
          "name": { "type": "string", "example": "Coffee" },
          "type": { "type": "string", "enum": ["expense", "income"], "example": "expense" },
          "icon": { "type": "string", "example": "coffee" },
          "color": { "type": "string", "example": "#FFAA00" },
          "sortOrder": { "type": "integer", "example": 1 }
        }
      },
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "餐饮" },
          "type": { "type": "string", "example": "expense" },
          "icon": { "type": "string", "example": "food" },
          "color": { "type": "string", "example": "#FF6B6B" },
          "sortOrder": { "type": "integer", "example": 1 },
          "createdAt": { "type": "string", "example": "2026-06-03 14:30:00" },
          "updatedAt": { "type": "string", "example": "2026-06-03 14:30:00" }
        }
      },
      "AccountCreateRequest": {
        "type": "object",
        "required": ["name", "type"],
        "properties": {
          "name": { "type": "string", "example": "招商银行" },
          "type": { "type": "string", "enum": ["cash", "bank_card", "wechat", "alipay", "credit_card", "other"], "example": "bank_card" },
          "balance": { "type": "number", "example": 10000.00 },
          "icon": { "type": "string", "example": "card" },
          "color": { "type": "string", "example": "#3377FF" },
          "sortOrder": { "type": "integer", "example": 2 }
        }
      },
      "AccountResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "现金" },
          "type": { "type": "string", "example": "cash" },
          "balance": { "type": "number", "example": 0.00 },
          "icon": { "type": "string", "example": "cash" },
          "color": { "type": "string", "example": "#2D3436" },
          "sortOrder": { "type": "integer", "example": 1 },
          "createdAt": { "type": "string", "example": "2026-06-03 14:30:00" },
          "updatedAt": { "type": "string", "example": "2026-06-03 14:30:00" }
        }
      },
      "BudgetCreateRequest": {
        "type": "object",
        "required": ["amount", "month"],
        "properties": {
          "categoryId": { "type": "integer", "nullable": true, "example": 1, "description": "分类ID，null表示总预算" },
          "amount": { "type": "number", "example": 1000.00, "description": "预算金额" },
          "month": { "type": "string", "example": "2026-06", "description": "月份 yyyy-MM" }
        }
      },
      "BudgetResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "categoryId": { "type": "integer", "nullable": true, "example": 1 },
          "categoryName": { "type": "string", "example": "餐饮" },
          "amount": { "type": "number", "example": 1000.00 },
          "month": { "type": "string", "example": "2026-06" },
          "createdAt": { "type": "string", "example": "2026-06-03 14:30:00" },
          "updatedAt": { "type": "string", "example": "2026-06-03 14:30:00" }
        }
      }
    },
    "parameters": {
      "X-App-Key": {
        "name": "X-App-Key",
        "in": "header",
        "required": true,
        "schema": { "type": "string", "default": "nl_flutter_2026" },
        "description": "应用密钥"
      },
      "TransactionId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" },
        "description": "记录ID"
      },
      "CategoryId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" },
        "description": "分类ID"
      },
      "AccountId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" },
        "description": "账户ID"
      },
      "BudgetId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" },
        "description": "预算ID"
      }
    }
  },
  "paths": {
    "/ping": {
      "get": {
        "tags": ["基础"],
        "summary": "测试后端连通性",
        "security": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "type": "string", "example": "server ok" } } }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["认证"],
        "summary": "用户注册",
        "security": [],
        "parameters": [{ "$ref": "#/components/parameters/X-App-Key" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/UserResponse" } } }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["认证"],
        "summary": "用户登录",
        "security": [],
        "parameters": [{ "$ref": "#/components/parameters/X-App-Key" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/LoginResponse" } } }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/user/me": {
      "get": {
        "tags": ["用户"],
        "summary": "获取当前用户信息",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/UserResponse" } } }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/transactions": {
      "get": {
        "tags": ["记账记录"],
        "summary": "查询记账列表",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["expense", "income"] }, "description": "类型筛选" },
          { "name": "categoryId", "in": "query", "schema": { "type": "integer" }, "description": "分类ID" },
          { "name": "accountId", "in": "query", "schema": { "type": "integer" }, "description": "账户ID" },
          { "name": "startDate", "in": "query", "schema": { "type": "string" }, "description": "开始时间 yyyy-MM-dd HH:mm:ss" },
          { "name": "endDate", "in": "query", "schema": { "type": "string" }, "description": "结束时间 yyyy-MM-dd HH:mm:ss" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/TransactionResponse" } } } }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["记账记录"],
        "summary": "新增记账记录",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransactionCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/TransactionResponse" } } }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{id}": {
      "get": {
        "tags": ["记账记录"],
        "summary": "查询记账详情",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TransactionId" }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/TransactionResponse" } } }
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["记账记录"],
        "summary": "修改记账记录",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TransactionId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransactionCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/TransactionResponse" } } }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["记账记录"],
        "summary": "删除记账记录",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TransactionId" }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "type": "boolean", "example": true } } }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": ["分类"],
        "summary": "查询分类列表",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["expense", "income"] }, "description": "类型筛选" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/CategoryResponse" } } } }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["分类"],
        "summary": "新增分类",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoryCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/CategoryResponse" } } }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/categories/{id}": {
      "put": {
        "tags": ["分类"],
        "summary": "修改分类",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/CategoryId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoryCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/CategoryResponse" } } }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["分类"],
        "summary": "删除分类",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/CategoryId" }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "type": "boolean", "example": true } } }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/accounts": {
      "get": {
        "tags": ["账户"],
        "summary": "查询账户列表",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/AccountResponse" } } } }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["账户"],
        "summary": "新增账户",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AccountCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/AccountResponse" } } }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{id}": {
      "put": {
        "tags": ["账户"],
        "summary": "修改账户",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AccountCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/AccountResponse" } } }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["账户"],
        "summary": "删除账户",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountId" }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "type": "boolean", "example": true } } }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/budgets": {
      "get": {
        "tags": ["预算"],
        "summary": "查询预算列表",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "month", "in": "query", "schema": { "type": "string" }, "description": "月份 yyyy-MM" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/BudgetResponse" } } } }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["预算"],
        "summary": "新增预算",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BudgetCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/BudgetResponse" } } }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/budgets/{id}": {
      "put": {
        "tags": ["预算"],
        "summary": "修改预算",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/BudgetId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BudgetCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/BudgetResponse" } } }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["预算"],
        "summary": "删除预算",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/BudgetId" }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "type": "boolean", "example": true } } }
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}