# server/src/main/java/com/aihy/nuclearledger/dto/account/AccountResponse.java
package com.aihy.nuclearledger.dto.account;

import com.fasterxml.jackson.annotation.JsonFormat;
import lombok.Data;

import java.math.BigDecimal;
import java.time.LocalDateTime;

@Data
public class AccountResponse {

    private Long id;
    private String name;
    private String type;
    private BigDecimal balance;
    private String icon;
    private String color;
    private Integer sortOrder;

    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
    private LocalDateTime createdAt;

    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
    private LocalDateTime updatedAt;

}